refactor: feed 컴포넌트 타입 정리 및 기타 코드 정리
